requirements_dev.txt
-r requirements.txt
setuptools>=74.0.0
pytest>=8.3.3
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
python-dotenv>=1.0.1